# Perform faster string termination check.