--- a/lib/source-map.js
+++ b/lib/source-map.js
@@ -88,7 +88,12 @@ class SourceMap {
       source = srcURL.removeFrom(source);
     }
 
-    if (this.content.mappings.length > 0 && !/[;,]$/.test(this.content.mappings)) {
+    const endingCharacters = {
+      ',': true,
+      ';': true,
+    };
+
+    if (this.content.mappings.length > 0 && !endingCharacters[this.content.mappings[this.content.mappings.length - 1]]) {
       this.content.mappings += ',';
     }