@gdborton 2018-03-01
Closing this out as I did some more testing and it appears whichever method is done first performs faster. Must be a JIT thing.