@ef4 2018-03-01
Thanks for following up. It takes a lot of effort to isolate real performance measurements from JIT and caching effects.